requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
httpx
python-dotenv